pyproject.toml
[tool.poetry]
name = "autorizaciones_salud"
version = "2.0.0"
description = "API for extracting information from images using Gemini"
authors = ["JAANDR7 <jaandr7@mapfre.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.0"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
uvicorn = "0.30.6"
google-genai = "1.24.0"
python-dotenv = "1.0.1"
requests = "2.32.3"
boto3 = "^1.35.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
